Clean up tests and make sure we remove unix sockets when we are done with them.
