+++ integration_tests/packet_test.go
@@ -5,36 +5,9 @@ import (
 
 	"encoding/hex"
 	"encoding/json"
-	"net"
 	"testing"
 	"time"
 )
-
-func WaitForSocket(p string) (net.Conn, error) {
-	timeout := time.After(time.Second)
-	for range time.Tick(10 * time.Millisecond) {
-		c, err := net.Dial("unix", "/tmp/unix")
-		if err == nil {
-			return c, err
-		}
-		select {
-		case <-timeout:
-			return c, err
-		default:
-		}
-	}
-	panic("Unreachable")
-}
-
-func WaitForPacket(c net.Conn, t *testing.T, s chan node.Packet) {
-	r := json.NewDecoder(c)
-	var p node.Packet
-	err := r.Decode(&p)
-	if err != nil {
-		t.Fatal(err)
-	}
-	s <- p
-}
 
 // This tests that packets sent over a packet connection fully transit the system.
 func TestPacket(t *testing.T) {
