+++ integration_tests/test_helpers.go
@@ -1,23 +1,28 @@
 package integration_tests
 
 import (
+	"github.com/AutoRoute/node"
+
+	"encoding/json"
 	"errors"
+	"fmt"
+	"net"
+	"testing"
 	"time"
 )
 
 func WaitForID(b AutoRouteBinary) (string, error) {
-	timeout := time.After(1 * time.Second)
+	timeout := time.After(2 * time.Second)
 	for range time.Tick(10 * time.Millisecond) {
+		id, err := b.GetID()
+		if err == nil {
+			return id, nil
+		}
 		select {
 		case <-timeout:
-			return "", errors.New("Timeout while waiting for id")
+			return "", errors.New(fmt.Sprint("Timeout while waiting for id:", err))
 		default:
 		}
-		id, err := b.GetID()
-		if err != nil {
-			continue
-		}
-		return id, nil
 	}
 	panic("unreachable")
 }
@@ -25,12 +30,6 @@ func WaitForID(b AutoRouteBinary) (string, error) {
 func WaitForConnection(b AutoRouteBinary, addr string) error {
 	stop := time.After(10 * time.Second)
 	for range time.Tick(10 * time.Millisecond) {
-		select {
-		case <-stop:
-			return errors.New("Timeout out while waiting for connection")
-		default:
-		}
-
 		connections, err := b.GetConnections()
 		if err != nil {
 			continue
@@ -40,6 +39,38 @@ func WaitForConnection(b AutoRouteBinary, addr string) error {
 				return nil
 			}
 		}
+
+		select {
+		case <-stop:
+			return errors.New("Timeout out while waiting for connection")
+		default:
+		}
 	}
 	panic("unreachable")
+}
+
+func WaitForSocket(p string) (net.Conn, error) {
+	timeout := time.After(time.Second)
+	for range time.Tick(10 * time.Millisecond) {
+		c, err := net.Dial("unix", "/tmp/unix")
+		if err == nil {
+			return c, err
+		}
+		select {
+		case <-timeout:
+			return c, err
+		default:
+		}
+	}
+	panic("Unreachable")
+}
+
+func WaitForPacket(c net.Conn, t *testing.T, s chan node.Packet) {
+	r := json.NewDecoder(c)
+	var p node.Packet
+	err := r.Decode(&p)
+	if err != nil {
+		t.Fatal(err)
+	}
+	s <- p
 }
